Config Fuzzer/Bisector
This tool makes it easy to search through config state-space with a minimal reproduction or test. It presents a similar interface to the config bisector by taking a test_function that should either raise on Exception or return False upon failure.

It has two entry points: fuzz_n_tuple, which tries every combination of n configs, and fuzz_with_bisect, which randomly flips configs and tries to find the minimal reproduction upon failure. fuzz_with_bisect is a much more efficient way to search the space, but fuzz_n_tuple can give you peace of mind that a new config will compose with every other config.

It's been used to find two bugs so far in the inductor config:

This PR also adds a bunch of missing types to the inductor config to get them to play nice with the fuzzer, so it can be a good forcing function for adding types to config.