[Redo] Set remote cache version and backend type once in compilation metrics
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141967

(Got reverted due to a silly bug, fixed now.)

This is causing FbFxGraphRemoteCache.init to no longer be idempotent, i.e. only safe to call once per compile. AOTAutogradCache initializes a new remote cache for the forward and the backward.
Technically, we could make AOTAutogradCache smart and globally thread through a single FbFxGraphRemoteCache everywhere. But there's no reason to do so, as this class is just the handle to access the cache. Plus, it's very brittle for FbFxGraphRemoteCache to not be safe to call multiple times
ghstack-source-id: 256244970
@exported-using-ghexport

Differential Revision: [D66701970](https://our.internmc.facebook.com/intern/diff/D66701970/)