Remove option for fork-based compile pool
Summary: This has been set to "subproc" for a while internally and externally, so we can remove and simplify some of the code. Note that there's no pressing need here -- just that since we've had internal outage with the legacy "fork" implementation, it doesn't seem helpful to leave it available. But if people aren't in the mood for this sort of cleanup, I won't be offended to abandon it.

Test Plan: CI

ghstack-source-id: ecf05d5bbc9276e1deae30774931d2d110c722cc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/142001